Move logic of currentProperties to the store as a getter
