+++ packages/plugins/Gfi/src/types.ts
@@ -43,6 +43,7 @@ export interface GfiState {
 
 export interface GfiGetters extends GfiState {
   afterLoadFunction: GfiAfterLoadFunction | null
+  currentProperties: GeoJsonProperties
   exportPropertyLayerKeys: Record<string, string>
   /** subset of layerKeys, where features' geometries are to be shown on map */
   geometryLayerKeys: string[]
